feat: remove agentImage/operatorImage fields, consolidate operator version to single source
Makefile ARGOCD_OPERATOR_IMAGE is now the single source of truth for the
ArgoCD operator version, injected at build time via ldflags. Removed
defaults.go, image env vars from addon pod, and agent image from ArgoCD
CR templates (operator auto-determines agent version). Consolidated
operator chart (embedded chart is source, hub chart is copy). Renamed
e2e tests to clearly distinguish basic-pull vs advanced-pull, merged
cleanup tests into main advanced-pull test lifecycle.
