+++ .github/workflows/e2e.yml
@@ -17,8 +17,8 @@ defaults:
     working-directory: go/src/open-cluster-management.io/argocd-pull-integration
 
 jobs:
-  e2e:
-    name: e2e
+  e2e-advanced-pull:
+    name: e2e-advanced-pull
     runs-on: ubuntu-latest
     steps:
       - name: checkout code
@@ -73,13 +73,13 @@ jobs:
         run: |
           kind load docker-image --name=cluster1 quay.io/open-cluster-management/argocd-pull-integration:latest
 
-      - name: Run e2e test
+      - name: Run advanced pull e2e test
         env:
           KUBECONFIG: /home/runner/.kube/config
-        run: make test-e2e
+        run: make test-e2e-advanced-pull
 
-  e2e-cleanup:
-    name: e2e-cleanup
+  e2e-advanced-pull-custom-namespace:
+    name: e2e-advanced-pull-custom-namespace
     runs-on: ubuntu-latest
     steps:
       - name: checkout code
@@ -132,13 +132,13 @@ jobs:
         run: |
           kind load docker-image --name=cluster1 quay.io/open-cluster-management/argocd-pull-integration:latest
 
-      - name: Run e2e cleanup test
+      - name: Run advanced pull custom namespace e2e test
         env:
           KUBECONFIG: /home/runner/.kube/config
-        run: make test-e2e-cleanup
+        run: make test-e2e-advanced-pull-custom-namespace
 
-  e2e-custom-namespace:
-    name: e2e-custom-namespace
+  e2e-basic-pull:
+    name: e2e-basic-pull
     runs-on: ubuntu-latest
     steps:
       - name: checkout code
@@ -191,66 +191,7 @@ jobs:
         run: |
           kind load docker-image --name=cluster1 quay.io/open-cluster-management/argocd-pull-integration:latest
 
-      - name: Run e2e custom namespace test
+      - name: Run basic pull e2e test
         env:
           KUBECONFIG: /home/runner/.kube/config
-        run: make test-e2e-custom-namespace
-
-  e2e-basic:
-    name: e2e-basic
-    runs-on: ubuntu-latest
-    steps:
-      - name: checkout code
-        uses: actions/checkout@v4
-        with:
-          fetch-depth: 1
-          path: go/src/open-cluster-management.io/argocd-pull-integration
-
-      - name: install Go
-        uses: actions/setup-go@v5
-        with:
-          go-version: ${{ env.GO_VERSION }}
-
-      - name: Add GOPATH/bin to PATH
-        run: echo "$GOPATH/bin" >> "$GITHUB_PATH"
-
-      - name: install imagebuilder
-        run: go install github.com/openshift/imagebuilder/cmd/imagebuilder@v1.2.3
-
-      - name: Set up Helm
-        uses: azure/setup-helm@v4
-        with:
-          version: v3.15.2
-
-      - name: Set up kubectl
-        uses: azure/setup-kubectl@v4
-        with:
-          version: v1.29.0
-
-      - name: build-images
-        run: make build-images
-
-      - name: setup kind (cluster1)
-        uses: engineerd/setup-kind@v0.5.0
-        with:
-          version: v0.14.0
-          name: cluster1
-
-      - name: setup kind (hub)
-        uses: engineerd/setup-kind@v0.5.0
-        with:
-          version: v0.14.0
-          name: hub
-
-      - name: Load image on the nodes of the hub
-        run: |
-          kind load docker-image --name=hub quay.io/open-cluster-management/argocd-pull-integration:latest
-
-      - name: Load image on the nodes of the cluster1
-        run: |
-          kind load docker-image --name=cluster1 quay.io/open-cluster-management/argocd-pull-integration:latest
-
-      - name: Run e2e basic test
-        env:
-          KUBECONFIG: /home/runner/.kube/config
-        run: make test-e2e-basic
+        run: make test-e2e-basic-pull
