+++ config/crd/bases/apps.open-cluster-management.io_gitopsclusters.yaml
@@ -50,9 +50,6 @@ spec:
                 description: ArgoCDAgentAddon defines the configuration for the ArgoCD
                   agent addon
                 properties:
-                  agentImage:
-                    description: AgentImage is the ArgoCD agent image to use
-                    type: string
                   agentNamespace:
                     description: AgentNamespace is the namespace where the ArgoCD
                       agent (ArgoCD CR) will be deployed on the managed cluster
@@ -60,9 +57,6 @@ spec:
                   mode:
                     default: managed
                     description: Mode defines the agent mode (managed or autonomous)
-                    type: string
-                  operatorImage:
-                    description: OperatorImage is the ArgoCD operator image to use
                     type: string
                   operatorNamespace:
                     description: OperatorNamespace is the namespace where the ArgoCD
