+++ internal/addon/addon_controller_test.go
@@ -32,44 +32,27 @@ func TestArgoCDAgentAddonReconciler(t *testing.T) {
 		name              string
 		interval          int
 		operatorImage     string
-		agentImage        string
 		serverAddress     string
 		serverPort        string
 		mode              string
-		uninstall         bool
 		wantIntervalValid bool
 	}{
 		{
 			name:              "default configuration",
 			interval:          60,
-			operatorImage:     "quay.io/operator:v1.0.0",
-			agentImage:        "quay.io/agent:v1.0.0",
+			operatorImage:     "quay.io/operator:latest",
 			serverAddress:     "argocd-server.argocd.svc",
 			serverPort:        "8080",
 			mode:              "managed",
-			uninstall:         false,
 			wantIntervalValid: true,
 		},
 		{
 			name:              "autonomous mode",
 			interval:          30,
-			operatorImage:     "quay.io/operator:v1.0.0",
-			agentImage:        "quay.io/agent:v1.0.0",
+			operatorImage:     "quay.io/operator:latest",
 			serverAddress:     "argocd-server.argocd.svc",
 			serverPort:        "8080",
 			mode:              "autonomous",
-			uninstall:         false,
-			wantIntervalValid: true,
-		},
-		{
-			name:              "uninstall mode",
-			interval:          60,
-			operatorImage:     "quay.io/operator:v1.0.0",
-			agentImage:        "quay.io/agent:v1.0.0",
-			serverAddress:     "argocd-server.argocd.svc",
-			serverPort:        "8080",
-			mode:              "managed",
-			uninstall:         true,
 			wantIntervalValid: true,
 		},
 	}
@@ -80,22 +63,17 @@ func TestArgoCDAgentAddonReconciler(t *testing.T) {
 				Client:                   fake.NewClientBuilder().WithScheme(s).Build(),
 				Scheme:                   s,
 				Interval:                 tt.interval,
-				ArgoCDOperatorImage:      tt.operatorImage,
-				ArgoCDAgentImage:         tt.agentImage,
+				OperatorImage:            tt.operatorImage,
 				ArgoCDAgentServerAddress: tt.serverAddress,
 				ArgoCDAgentServerPort:    tt.serverPort,
 				ArgoCDAgentMode:          tt.mode,
 			}
 
-			// Verify the reconciler is properly configured
 			if r.Interval != tt.interval {
 				t.Errorf("Interval = %v, want %v", r.Interval, tt.interval)
 			}
-			if r.ArgoCDOperatorImage != tt.operatorImage {
-				t.Errorf("OperatorImage = %v, want %v", r.ArgoCDOperatorImage, tt.operatorImage)
-			}
-			if r.ArgoCDAgentImage != tt.agentImage {
-				t.Errorf("AgentImage = %v, want %v", r.ArgoCDAgentImage, tt.agentImage)
+			if r.OperatorImage != tt.operatorImage {
+				t.Errorf("OperatorImage = %v, want %v", r.OperatorImage, tt.operatorImage)
 			}
 			if r.ArgoCDAgentServerAddress != tt.serverAddress {
 				t.Errorf("ServerAddress = %v, want %v", r.ArgoCDAgentServerAddress, tt.serverAddress)
