+++ internal/addon/charts/argocd-agent-addon/templates/argocd.yaml
@@ -12,7 +12,6 @@ spec:
       creds: "mtls:open-cluster-management:cluster:([^:]+):addon:argocd-agent"
       logLevel: "info"
       logFormat: "text"
-      image: {{ .Values.global.argoCDAgent.image }}:{{ .Values.global.argoCDAgent.tag }}
       client:
         principalServerAddress: {{ .Values.global.argoCDAgent.serverAddress }}
         principalServerPort: "{{ .Values.global.argoCDAgent.serverPort }}"
