+++ internal/addon/charts/argocd-agent-addon/templates/operator.yaml
@@ -405,11 +405,13 @@ spec:
         env:
         - name: ARGOCD_CLUSTER_CONFIG_NAMESPACES
           value: "*"
+        - name: ARGOCD_PRINCIPAL_TLS_SERVER_ALLOW_GENERATE
+          value: "false"
         - name: WATCH_NAMESPACE
           valueFrom:
             fieldRef:
               fieldPath: metadata.annotations['olm.targetNamespaces']
-        image: {{ .Values.operatorImage }}:{{ .Values.operatorImageTag }}
+        image: "{{ .Values.operatorImage }}:{{ .Values.operatorImageTag }}"
         livenessProbe:
           httpGet:
             path: /healthz
